firebase/beta/requirements.txt
firebase_functions
google-genai
Pillow
legacy-cgi